data_preprocessing/requirements.txt
beautifulsoup4==4.10.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cssselect==1.1.0
cymem==2.0.6
dnspython==2.1.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
idna==3.3
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
lxml==4.6.4
MarkupSafe==2.0.1
mongoengine==0.23.1
murmurhash==1.0.6
newspaper3k==0.2.8
nltk==3.6.5
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pathy==0.6.1
Pillow==8.4.0
preshed==3.0.6
pydantic==1.8.2
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
scikit-learn==1.0.1
scipy==1.7.3
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
sonopy==0.1.2
soupsieve==2.3.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
textblob==0.17.1
thinc==8.0.13
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.0
Unidecode==1.3.2
urllib3==1.26.7
vader==0.0.3
vaderSentiment==3.3.2
wasabi==0.8.2
xlrd==2.0.1
xmltodict==0.12.0
